requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django
django-filter==24.1
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
idna==3.6
phonenumbers==8.13.32
pillow==10.2.0
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.2.1
